+++ packages/sdk/src/utils/index.ts
@@ -9,4 +9,5 @@
  *
  */
 
+export * from "./SessionContext";
 export * from "./tools/SearchParser";